+++ cognite_toolkit/_cdf_tk/apps/__init__.py
@@ -1,6 +1,7 @@
+from ._core_app import CoreApp
 from ._landing_app import LandingApp
 from ._modules_app import ModulesApp
 from ._repo_app import RepoApp
 from ._run import RunApp
 
-__all__ = ["LandingApp", "ModulesApp", "RunApp", "RepoApp"]
+__all__ = ["CoreApp", "LandingApp", "ModulesApp", "RunApp", "RepoApp"]
